requirements.txt
asgiref==3.9.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.6
django-allauth==65.12.0
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.16.1
drf-yasg==1.21.7
factory_boy==3.3.3
Faker==37.11.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.8.0
pytest==8.4.2
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
django-allauth==65.12.0

